chore: move trpc initialization file to library
