+++ src/modules/agents/ui/views/agent-view.tsx
@@ -15,9 +15,9 @@ import { GeneratedAvatar } from '@/components/generated-avatar';
 import { LoadingState } from '@/components/loading-state';
 import { Badge } from '@/components/ui/badge';
 import { useConfirm } from '@/hooks/use-confirm';
+import { useTRPC } from '@/lib/trpc';
 import { AgentHeader } from '@/modules/agents/ui/components/agent-header';
 import { UpdateAgentDialog } from '@/modules/agents/ui/components/update-agent-dialog';
-import { useTRPC } from '@/utils/trpc';
 
 interface Props {
   agentId: string;
