+++ src/modules/meetings/ui/components/transcript.tsx
@@ -7,7 +7,7 @@ import Highlighter from 'react-highlight-words';
 import { Avatar, AvatarImage } from '@/components/ui/avatar';
 import { Input } from '@/components/ui/input';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import { useTRPC } from '@/utils/trpc';
+import { useTRPC } from '@/lib/trpc';
 
 interface Props {
   meetingId: string;
