+++ src/modules/meetings/ui/views/meeting-view.tsx
@@ -12,6 +12,7 @@ import { toast } from 'sonner';
 import { ErrorState } from '@/components/error-state';
 import { LoadingState } from '@/components/loading-state';
 import { useConfirm } from '@/hooks/use-confirm';
+import { useTRPC } from '@/lib/trpc';
 import { ActiveState } from '@/modules/meetings/ui/components/active-state';
 import { CancelledState } from '@/modules/meetings/ui/components/cancelled-state';
 import { CompletedState } from '@/modules/meetings/ui/components/completed-state';
@@ -19,7 +20,6 @@ import { MeetingHeader } from '@/modules/meetings/ui/components/meeting-header';
 import { ProcessingState } from '@/modules/meetings/ui/components/processing-state';
 import { UpcomingState } from '@/modules/meetings/ui/components/upcoming-state';
 import { UpdateMeetingDialog } from '@/modules/meetings/ui/components/update-meeting-dialog';
-import { useTRPC } from '@/utils/trpc';
 
 interface Props {
   meetingId: string;
