+++ src/trpc/client.tsx
@@ -7,9 +7,9 @@ import { useState } from 'react';
 import superjson from 'superjson';
 
 import { API_URL } from '@/lib/env';
+import { TRPCProvider } from '@/lib/trpc';
 import { makeQueryClient } from '@/trpc/query-client';
 import type { AppRouter } from '@/trpc/routers/_app';
-import { TRPCProvider } from '@/utils/trpc';
 
 let browserQueryClient: QueryClient | undefined = undefined;
